api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.30
python-dateutil==2.9.0
pydantic==2.7.4
pytest==8.3.2
httpx==0.27.2
python-json-logger==2.0.7
prometheus-client==0.20.0
asyncpg==0.29.0
